Expressjs initial setup

+++ code/backend/vagrant/config/bootstrap.sh
@@ -1,21 +1,25 @@
 #!/bin/sh -e
 
+export DEBIAN_FRONTEND=noninteractive
+
 # Update apt-get
 apt-get update
+apt-get -y upgrade
+
+PROVISIONED_ON=/etc/vm_provision_on_timestamp
+if [ -f "$PROVISIONED_ON" ]
+then
+  echo "VM was already provisioned at: $(cat $PROVISIONED_ON)"
+  echo "To run system updates manually login via 'vagrant ssh' and run 'apt-get update && apt-get upgrade'"
+  echo ""
+  exit
+fi
 
 # Install Curl
 apt-get install -y curl
 
 # Git
 apt-get install -y git
-
-# Install Node.js and NPM
-add-apt-repository ppa:chris-lea/node.js
-apt-get update
-apt-get install -y nodejs
-
-# Install Grunt
-npm install -g grunt-cli
 
 ################################################
 ############## PostgreSQL section ##############
@@ -31,45 +35,6 @@ APP_DB_NAME=$APP_DB_USER
 # Edit the following to change the version of PostgreSQL that is installed
 PG_VERSION=9.4
 
-###########################################################
-# Changes below this line are probably not necessary
-###########################################################
-print_db_usage () {
-  echo "Your PostgreSQL database has been setup and can be accessed on your local machine on the forwarded port (default: 15432)"
-  echo "  Host: localhost"
-  echo "  Port: 15432"
-  echo "  Database: $APP_DB_NAME"
-  echo "  Username: $APP_DB_USER"
-  echo "  Password: $APP_DB_PASS"
-  echo ""
-  echo "Admin access to postgres user via VM:"
-  echo "  vagrant ssh"
-  echo "  sudo su - postgres"
-  echo ""
-  echo "psql access to app database user via VM:"
-  echo "  vagrant ssh"
-  echo "  sudo su - postgres"
-  echo "  PGUSER=$APP_DB_USER PGPASSWORD=$APP_DB_PASS psql -h localhost $APP_DB_NAME"
-  echo ""
-  echo "Env variable for application development:"
-  echo "  DATABASE_URL=jdbc:postgresql://$APP_DB_USER:$APP_DB_PASS@localhost:15432/$APP_DB_NAME"
-  echo ""
-  echo "Local command to access the database via psql:"
-  echo "  PGUSER=$APP_DB_USER PGPASSWORD=$APP_DB_PASS psql -h localhost -p 15432 $APP_DB_NAME"
-}
-
-export DEBIAN_FRONTEND=noninteractive
-
-PROVISIONED_ON=/etc/vm_provision_on_timestamp
-if [ -f "$PROVISIONED_ON" ]
-then
-  echo "VM was already provisioned at: $(cat $PROVISIONED_ON)"
-  echo "To run system updates manually login via 'vagrant ssh' and run 'apt-get update && apt-get upgrade'"
-  echo ""
-  print_db_usage
-  exit
-fi
-
 PG_REPO_APT_SOURCE=/etc/apt/sources.list.d/pgdg.list
 if [ ! -f "$PG_REPO_APT_SOURCE" ]
 then
@@ -81,8 +46,6 @@ then
 fi
 
 # Update package list and upgrade all packages
-apt-get update
-apt-get -y upgrade
 
 apt-get -y install "postgresql-$PG_VERSION" "postgresql-contrib-$PG_VERSION"
 
@@ -114,12 +77,47 @@ CREATE DATABASE $APP_DB_NAME WITH OWNER=$APP_DB_USER
                                   TEMPLATE=template0;
 EOF
 
-# Tag the provision time:
-date > "$PROVISIONED_ON"
-
 echo "Successfully created PostgreSQL dev virtual machine."
 echo ""
-print_db_usage
+echo "Your PostgreSQL database has been setup and can be accessed on your local machine on the forwarded port (default: 15432)"
+echo "  Host: localhost"
+echo "  Port: 15432"
+echo "  Database: $APP_DB_NAME"
+echo "  Username: $APP_DB_USER"
+echo "  Password: $APP_DB_PASS"
+echo ""
+echo "Admin access to postgres user via VM:"
+echo "  vagrant ssh"
+echo "  sudo su - postgres"
+echo ""
+echo "psql access to app database user via VM:"
+echo "  vagrant ssh"
+echo "  sudo su - postgres"
+echo "  PGUSER=$APP_DB_USER PGPASSWORD=$APP_DB_PASS psql -h localhost $APP_DB_NAME"
+echo ""
+echo "Env variable for application development:"
+echo "  DATABASE_URL=jdbc:postgresql://$APP_DB_USER:$APP_DB_PASS@localhost:15432/$APP_DB_NAME"
+echo ""
+echo "Local command to access the database via psql:"
+echo "  PGUSER=$APP_DB_USER PGPASSWORD=$APP_DB_PASS psql -h localhost -p 15432 $APP_DB_NAME"
+echo ""
+
+# Install Node.js and NPM
+add-apt-repository ppa:chris-lea/node.js
+apt-get update
+apt-get install -y nodejs
+
+# Install Grunt
+npm install -g grunt-cli
+
+# Install the NPM dependencies
+cd /seubiu && npm install
+
+#Run the server
+cd /seubiu && DEBUG=:* npm start
 
 echo ""
 echo "The development environment was created successfully."
+
+# Tag the provision time:
+date > "$PROVISIONED_ON"
